requirements.txt
pydantic>=2.7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
qdrant-client==1.16.0
llama-index==0.10.45.post1
llama-index-vector-stores-qdrant==0.2.17
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
pdfplumber>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0